requirements.txt
-e .
aws_cdk.aws_iam
aws_cdk.aws_ec2
aws_cdk.aws_redshift
aws_cdk.aws_s3